android: Add --abi option to android/run-checks.sh
This patch adds an option to android/run-checks.sh to specify
the target ABI. For example, by using the following line:

  android/run-checks.sh --abi=x86 --no-device

One can check the Android/x86 build of Google Breakpad (which
still fails for reasons that will be fixed in later patches).

Another use is to force the 'armeabi' ABI even when an
ARMv7-A device is connected (which uses the 'armeabi-v7a'
by default, but supports 'armeabi' as well).
Review URL: https://breakpad.appspot.com/424002